Split up store into separate files for state, mutations, actions, etc.
Also renames the model classes to avoid using reserved names
