+++ resources/models/Suggestion.js
@@ -6,7 +6,7 @@
  * @param {string} text The label text
  * @param {string} wikidataId The wikidata ID
  */
-module.exports = function SuggestionData( text, wikidataId ) {
+module.exports = function MvSuggestion( text, wikidataId ) {
 	this.text = text;
 	this.wikidataId = wikidataId;
 	this.confirmed = false;
